+++ src/pystra/distributions/shifted_exponential.py
@@ -3,7 +3,7 @@
 
 from scipy.stats import expon
 
-from .distribution import Distribution
+from .distribution import Distribution, _uses_native_parameters
 
 __all__ = ["ShiftedExponential"]
 
@@ -13,19 +13,22 @@ class ShiftedExponential(Distribution):
 
     :Attributes:
         - name (str):         Name of the random variable\n
-        - mean (float):       Mean or lamb\n
-        - std (float):       Standard deviation or x_zero\n
-        - input_type (any):   Change meaning of mean and std\n
+        - mean (float):       Mean\n
+        - std (float):       Standard deviation\n
+        - rate (float): Rate, given instead of mean and std\n
+        - shift (float): Lower bound, given instead of mean and std\n
         - start_point (float): Start point for seach\n
     """
 
-    def __init__(self, name, mean, std, input_type=None, start_point=None):
-        if input_type is None:
+    def __init__(
+        self, name, mean=None, std=None, *, rate=None, shift=None, start_point=None
+    ):
+        if _uses_native_parameters(self, mean, std, rate=rate, shift=shift):
+            lamb = rate
+            x_zero = shift
+        else:
             x_zero = mean - std
             lamb = 1 / std
-        else:
-            lamb = mean
-            x_zero = std
 
         # use scipy to do the heavy lifting
         self.dist_obj = expon(loc=x_zero, scale=1 / lamb)