+++ src/pystra/distributions/shifted_rayleigh.py
@@ -4,7 +4,7 @@
 import numpy as np
 from scipy.stats import rayleigh
 
-from .distribution import Distribution
+from .distribution import Distribution, _uses_native_parameters
 
 __all__ = ["ShiftedRayleigh"]
 
@@ -14,19 +14,22 @@ class ShiftedRayleigh(Distribution):
 
     :Attributes:
       - name (str):   Name of the random variable\n
-      - mean (float): Mean or a\n
-      - std (float): Standard deviation or x_zero\n
-      - input_type (any): Change meaning of mean and std\n
+      - mean (float): Mean\n
+      - std (float): Standard deviation\n
+      - scale (float): Scale, given instead of mean and std\n
+      - shift (float): Lower bound, given instead of mean and std\n
       - start_point (float): Start point for seach\n
     """
 
-    def __init__(self, name, mean, std, input_type=None, start_point=None):
-        if input_type is None:
+    def __init__(
+        self, name, mean=None, std=None, *, scale=None, shift=None, start_point=None
+    ):
+        if _uses_native_parameters(self, mean, std, scale=scale, shift=shift):
+            a = scale
+            x_zero = shift
+        else:
             a = std / ((2 - np.pi * 0.5) ** 0.5)
             x_zero = mean - std * (np.pi / (4 - np.pi)) ** 0.5
-        else:
-            a = mean
-            x_zero = std
 
         # use scipy to do the heavy lifting
         self.dist_obj = rayleigh(loc=x_zero, scale=a)